bug resuelto de los detaches

+++ app/Http/Controllers/ControladorEvidencias.php
@@ -202,8 +202,10 @@ class ControladorEvidencias extends Controller
     public function destroy($id)
     {
         $evidencia = Evidencia::findOrFail($id);
-                
-        $evidencia->planes()->detach($evidencia->planes[0]->id);
+            
+        if(count($evidencia->planes)> 0) {
+            $evidencia->planes()->detach($evidencia->planes[0]->id);
+        }
 
         $evidencia->delete();
         return redirect()->route('evidencias.index');
